# finance/lib/api-zod/src/generated/types/contractResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Financial calculation API for construction consulting
 * OpenAPI spec version: 0.1.0
 */

export interface ContractResult {
  revenue: number;
  kickback: number;
  laborCost: number;
  signingCost: number;
  fixedCost: number;
  taxableIncome: number;
  corporateTax: number;
  vatOut: number;
  vatIn: number;
  vatDue: number;
  netProfit: number;
  netProfitMonthly: number;
  netMargin: number;
  cashFlowBeforeTax: number;
  breakEvenRevenue: number;
  technicianWageRatio: number;
  isViable: boolean;
  /** Max total wage fund for Type 2 (revenue * maxWageFundRate); 0 for Type 1 */
  wageFundLimit: number;
  /** Actual total labor cost (director + accountant + technicians with insurance) */
  wageFundUsed: number;
  /** wageFundLimit - wageFundUsed; negative means over budget */
  wageFundRemaining: number;
  /** Max technicians that fit within the wage fund after director + accountant */
  maxTechniciansAllowed: number;
  /** True when current labor cost exceeds the wage fund limit (Type 2 only) */
  wageFundExceeded: boolean;
}